@@ -2,57 +2,41 @@
   "name": "monitoror",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "test": "npm-run-all test:unit",
-    "lint": "vue-cli-service lint",
-    "test:unit": "vue-cli-service test:unit"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint ."
   },
   "main": "index.js",
   "dependencies": {
-    "@babel/core": "^7.8.3",
     "axios": "^0.19.2",
-    "core-js": "^3.6.4",
     "date-fns": "^2.9.0",
     "lodash-es": "^4.17.15",
     "string-similarity": "^4.0.1",
     "ts-md5": "^1.2.7",
-    "vue": "^2.6.11",
-    "vue-class-component": "^7.2.2",
-    "vue-property-decorator": "^8.3.0",
-    "vuex": "^3.1.2"
+    "vue": "^3.0.2",
+    "vuex": "^4.0.0-beta.4"
   },
   "devDependencies": {
-    "@types/jest": "^25.1.2",
     "@types/lodash-es": "^4.17.3",
+    "@types/lodash.template": "^4.5.0",
     "@types/string-similarity": "^3.0.0",
-    "@typescript-eslint/eslint-plugin": "^2.17.0",
-    "@typescript-eslint/parser": "^2.17.0",
-    "@vue/cli-plugin-babel": "^4.1.2",
-    "@vue/cli-plugin-eslint": "^4.1.2",
-    "@vue/cli-plugin-typescript": "^4.1.2",
-    "@vue/cli-plugin-unit-jest": "^4.1.2",
-    "@vue/cli-service": "^4.1.2",
-    "@vue/eslint-config-standard": "^5.1.0",
-    "@vue/eslint-config-typescript": "^5.0.1",
-    "@vue/test-utils": "^1.0.0-beta.31",
-    "babel-eslint": "^11.0.0-beta.2",
-    "babel-jest": "^25.1.0",
-    "babel-loader": "^8.0.6",
-    "eslint": "^6.8.0",
-    "eslint-plugin-import": "^2.20.0",
-    "eslint-plugin-node": "^11.0.0",
+    "@typescript-eslint/eslint-plugin": "^4.6.0",
+    "@typescript-eslint/parser": "^4.6.0",
+    "@vue/compiler-sfc": "^3.0.2",
+    "@vue/eslint-config-standard": "^5.1.2",
+    "@vue/eslint-config-typescript": "^7.0.0",
+    "autoprefixer": "^9.8.6",
+    "eslint": "^7.12.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "eslint-plugin-vue": "^6.1.2",
-    "jest": "^25.1.0",
-    "node-sass": "^4.13.1",
-    "npm-run-all": "^4.1.5",
-    "sass-loader": "^8.0.2",
-    "ts-jest": "^25.0.0",
-    "typescript": "^3.7.5",
-    "vue-template-compiler": "^2.6.11",
-    "webpack": "^4.41.5"
+    "eslint-plugin-standard": "^4.0.2",
+    "eslint-plugin-vue": "^7.1.0",
+    "lodash.template": "^4.5.0",
+    "sass": "^1.27.0",
+    "typescript": "^4.0.3",
+    "vite": "^1.0.0-rc.7",
+    "vite-tsconfig-paths": "^1.1.0"
   },
   "browserslist": [
     "defaults",
