@@ -9,13 +9,8 @@
 
       <div class="c-monitoror-tile--build-info" v-if="branch || buildId">
         <template v-if="branch">
-          <svg v-if="mergeRequest" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" fill="currentColor">
-            <!-- Merge request icon -->
-            <path fill-rule="evenodd" d="M7.177 3.073L9.573.677A.25.25 0 0110 .854v4.792a.25.25 0 01-.427.177L7.177 3.427a.25.25 0 010-.354zM3.75 2.5a.75.75 0 100 1.5.75.75 0 000-1.5zm-2.25.75a2.25 2.25 0 113 2.122v5.256a2.251 2.251 0 11-1.5 0V5.372A2.25 2.25 0 011.5 3.25zM11 2.5h-1V4h1a1 1 0 011 1v5.628a2.251 2.251 0 101.5 0V5A2.5 2.5 0 0011 2.5zm1 10.25a.75.75 0 111.5 0 .75.75 0 01-1.5 0zM3.75 12a.75.75 0 100 1.5.75.75 0 000-1.5z"/>
-          </svg>
-          <svg v-else xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" fill="currentColor">
-            <!-- Branch icon -->
-            <path fill-rule="evenodd" d="M11.75 2.5a.75.75 0 100 1.5.75.75 0 000-1.5zm-2.25.75a2.25 2.25 0 113 2.122V6A2.5 2.5 0 0110 8.5H6a1 1 0 00-1 1v1.128a2.251 2.251 0 11-1.5 0V5.372a2.25 2.25 0 111.5 0v1.836A2.492 2.492 0 016 7h4a1 1 0 001-1v-.628A2.25 2.25 0 019.5 3.25zM4.25 12a.75.75 0 100 1.5.75.75 0 000-1.5zM3.5 3.25a.75.75 0 111.5 0 .75.75 0 01-1.5 0z"/>
+          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" fill="currentColor">
+            <use :xlink:href="'./icons.svg#' + (mergeRequest ? 'merge-request' : 'branch')"/>
           </svg>
           {{ branch }}
         </template>
@@ -29,8 +24,8 @@
         {{ message }}
       </div>
 
-      <div class="c-monitoror-tile--value" v-if="displayedValue">
-        {{ displayedValue }}
+      <div class="c-monitoror-tile--value" v-if="displayedMetric">
+        {{ displayedMetric }}
       </div>
 
       <div class="c-monitoror-tile--sub-tiles" v-if="isGroup">
@@ -67,150 +62,183 @@
 </template>
 
 <script lang="ts">
-  import {Component} from 'vue-property-decorator'
+import useTileValues from '@/composables/useTileMetrics'
+import {defineComponent, computed} from 'vue'
+import {useStore} from 'vuex'
 
-  import DISPLAYABLE_SUBTILE_STATUS from '@/constants/displayableSubtileStatus'
-  import TileStatus from '@/enums/tileStatus'
-  import TileType from '@/enums/tileType'
-  import TileValueUnit from '@/enums/tileValueUnit'
-  import TileConfig from '@/interfaces/tileConfig'
+import DISPLAYABLE_SUBTILE_STATUS from '@/constants/displayableSubtileStatus'
+import useTileCommons from '@/composables/useTileCommons'
+import TileStatus from '@/enums/tileStatus'
+import TileType from '@/enums/tileType'
+import TileConfig from '@/types/tileConfig'
 
-  import AbstractMonitororTile from '@/classes/monitororTile'
-  import MonitororSubTile from '@/components/SubTile.vue'
-  import MonitororTileIcon from '@/components/TileIcon.vue'
-  import TileValue from '@/interfaces/tileValue'
+import MonitororSubTile from '@/components/SubTile.vue'
+import MonitororTileIcon from '@/components/TileIcon.vue'
 
-  @Component({
-    components: {
-      MonitororSubTile,
-      MonitororTileIcon,
-    },
-  })
-  export default class MonitororTile extends AbstractMonitororTile {
+export default defineComponent({
+  name: 'MonitororTile',
+  components: {
+    MonitororSubTile,
+    MonitororTileIcon,
+  },
+  props: {
+    config: {
+      type: Object as () => TileConfig,
+      required: true
+    }
+  },
+  setup: function (props) {
+    const {
+      state,
+      theme,
+      type,
 
-    /*
-     * Computed
-     */
+      // status
+      status,
+      isQueued,
+      isRunning,
+      isSucceeded,
+      isFailed,
+      isWarning,
 
-    get classes() {
+      // content
+      label,
+
+      // build
+      build,
+      branch,
+      mergeRequest,
+      mergeRequestLabelPrefix,
+      progressTime,
+      progressBarStyle,
+      isOvertime,
+      finishedSince,
+      author,
+      showAuthor,
+    } = useTileCommons(props.config)
+
+    const {
+      displayedMetric,
+    } = useTileValues(state)
+
+    const store = useStore()
+
+    const classes = computed((): Record<string, boolean | string> => {
       return {
-        ['c-monitoror-tile__theme-' + this.theme]: true,
-        'c-monitoror-tile__empty': this.isEmpty,
-        'c-monitoror-tile__group': this.isGroup,
-        'c-monitoror-tile__status-succeeded': this.isSucceeded,
-        'c-monitoror-tile__status-failed': this.isFailed,
-        'c-monitoror-tile__status-warning': this.isWarning,
-        'c-monitoror-tile__status-running': this.isRunning,
-        'c-monitoror-tile__status-queued': this.isQueued,
-        'c-monitoror-tile__status-canceled': this.status === TileStatus.Canceled,
-        'c-monitoror-tile__status-action-required': this.status === TileStatus.ActionRequired,
+        ['c-monitoror-tile__theme-' + theme.value]: true,
+        'c-monitoror-tile__empty': isEmpty.value,
+        'c-monitoror-tile__group': isGroup.value,
+        'c-monitoror-tile__status-succeeded': isSucceeded.value,
+        'c-monitoror-tile__status-failed': isFailed.value,
+        'c-monitoror-tile__status-warning': isWarning.value,
+        'c-monitoror-tile__status-running': isRunning.value,
+        'c-monitoror-tile__status-queued': isQueued.value,
+        'c-monitoror-tile__status-canceled': status.value === TileStatus.Canceled,
+        'c-monitoror-tile__status-action-required': status.value === TileStatus.ActionRequired,
       }
-    }
+    })
 
-    get styles() {
-      const styles = {
-        'grid-column': `auto / span ${this.columnSpan}`,
-        'grid-row': `auto / span ${this.rowSpan}`,
-        '--row-span': this.rowSpan,
+    const styles = computed((): Record<string, string | number> => {
+      return {
+        'grid-column': `auto / span ${columnSpan.value}`,
+        'grid-row': `auto / span ${rowSpan.value}`,
+        '--row-span': rowSpan.value,
       }
+    })
 
-      return styles
-    }
+    const isEmpty = computed((): boolean => {
+      return type.value === TileType.Empty
+    })
 
-    get isEmpty(): boolean {
-      return this.type === TileType.Empty
-    }
+    const isGroup = computed((): boolean => {
+      return type.value === TileType.Group
+    })
 
-    get isGroup(): boolean {
-      return this.type === TileType.Group
-    }
+    const columnSpan = computed((): number => {
+      return props.config.columnSpan || 1
+    })
 
-    get columnSpan(): number {
-      return this.config.columnSpan || 1
-    }
+    const rowSpan = computed((): number => {
+      return props.config.rowSpan || 1
+    })
 
-    get rowSpan(): number {
-      return this.config.rowSpan || 1
-    }
-
-    get displayedSubTiles(): TileConfig[] | undefined {
-      if (!this.config.tiles) {
+    const displayedSubTiles = computed((): TileConfig[] | undefined => {
+      if (!props.config.tiles) {
         return
       }
 
-      const displayedSubTiles = this.config.tiles.filter((subTile) => {
-        if (!this.$store.state.tilesState.hasOwnProperty(subTile.stateKey)) {
+      const displayedSubTiles = props.config.tiles.filter((subTile) => {
+        if (!Object.keys(store.state.tilesState).includes(subTile.stateKey)) {
           return false
         }
 
-        return DISPLAYABLE_SUBTILE_STATUS.includes(this.$store.state.tilesState[subTile.stateKey].status)
+        return DISPLAYABLE_SUBTILE_STATUS.includes(store.state.tilesState[subTile.stateKey].status)
       })
 
       return displayedSubTiles
-    }
+    })
 
-    get value(): TileValue | undefined {
-      if (this.state === undefined) {
+    const message = computed((): string | undefined => {
+      if (state.value === undefined) {
         return
       }
 
-      return this.state.value
-    }
+      return state.value.message
+    })
 
-    get message(): string | undefined {
-      if (this.state === undefined) {
+    const buildId = computed((): string | undefined => {
+      if (build.value === undefined) {
         return
       }
 
-      return this.state.message
+      return build.value.id
+    })
+
+    return {
+      // attributes
+      classes,
+      styles,
+
+      // type
+      type,
+      isEmpty,
+      isGroup,
+
+      // status
+      status,
+      isQueued,
+      isRunning,
+      isSucceeded,
+      isFailed,
+      isWarning,
+
+      // layout
+      columnSpan,
+      rowSpan,
+
+      // content
+      label,
+      message,
+      displayedSubTiles,
+
+      // build
+      build,
+      buildId,
+      branch,
+      mergeRequest,
+      mergeRequestLabelPrefix,
+      progressTime,
+      progressBarStyle,
+      isOvertime,
+      finishedSince,
+      author,
+      showAuthor,
+
+      // metrics
+      displayedMetric,
     }
-
-    get buildId(): string | undefined {
-      if (this.build === undefined) {
-        return
-      }
-
-      return this.build.id
-    }
-
-    get unit(): TileValueUnit {
-      if (this.value === undefined) {
-        return TileValueUnit.Raw
-      }
-
-      return this.value.unit as TileValueUnit
-    }
-
-    get values(): string[] | undefined {
-      if (this.value === undefined) {
-        return
-      }
-
-      return this.value.values
-    }
-
-    get displayedValue(): string | undefined {
-      if (this.values === undefined) {
-        return
-      }
-
-      const UNIT_DISPLAY = {
-        [TileValueUnit.Millisecond]: 'ms',
-        [TileValueUnit.Ratio]: '%',
-        [TileValueUnit.Number]: '',
-        [TileValueUnit.Raw]: '',
-      }
-
-      let value = this.values[this.values.length - 1]
-      if (this.unit === TileValueUnit.Millisecond) {
-        value = Math.round(parseFloat(value)).toString()
-      } else if (this.unit === TileValueUnit.Ratio) {
-        value = (parseFloat(value) * 100).toFixed(2).toString()
-      }
-
-      return value + UNIT_DISPLAY[this.unit]
-    }
-  }
+  },
+})
 </script>
 
 <style lang="scss">
