@@ -1,135 +1,73 @@
 <template>
   <div class="c-monitoror-tile-icon">
-    <svg v-if="isGroup" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-      <path d="M21.41 17.91l-9 4A1.06 1.06 0 0 1 12 22a1 1 0 0 1-.41-.09l-9-4a1 1 0 0 1 .82-1.82L12 19.91l8.59-3.82a1 1 0 0 1 .82 1.82zm.5-6.32a1 1 0 0 0-1.32-.5L12 14.91l-8.59-3.82a1 1 0 0 0-.82 1.82l9 4A1 1 0 0 0 12 17a1.06 1.06 0 0 0 .4-.08l9-4a1 1 0 0 0 .51-1.33zm-.5-3.68l-9 4A1.06 1.06 0 0 1 12 12a1 1 0 0 1-.41-.09l-9-4a1 1 0 0 1 0-1.82l9-4A1.06 1.06 0 0 1 12 2a1 1 0 0 1 .41.09l9 4a.94.94 0 0 1 .5.5 1 1 0 0 1-.5 1.32zM18.54 7L12 4.09 5.46 7 12 9.91z"
-        stroke="var(--tile-background)" stroke-width="0.3" fill="currentColor" style="transform: translate(-2px, 2px)"></path>
-    </svg>
-
-    <svg v-else-if="isHttp" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 128 128" fill="currentColor">
-      <path d="M-.352 82.419V45.581h7.438V60.08H21.66V45.581h7.438v36.838H21.66V66.312H7.086v16.107H-.352zM45.079 82.419V51.813H34.148v-6.232h29.274v6.232H52.517v30.606h-7.438zM76.514 82.419V51.813H65.583v-6.232h29.274v6.232H83.951v30.606h-7.437zM99.656 82.419V45.581h11.936c4.523 0 7.471.185 8.845.553 2.11.553 3.878 1.755 5.302 3.606 1.423 1.851 2.135 4.242 2.135 7.174 0 2.262-.41 4.163-1.231 5.704-.821 1.541-1.864 2.751-3.128 3.631-1.265.879-2.551 1.462-3.857 1.746-1.776.352-4.347.528-7.715.528h-4.849v13.896h-7.438zm7.438-30.606v10.453h4.071c2.931 0 4.891-.192 5.879-.578.989-.385 1.763-.988 2.325-1.809.561-.821.841-1.776.841-2.864 0-1.341-.393-2.446-1.181-3.317-.787-.871-1.784-1.416-2.99-1.634-.888-.167-2.672-.251-5.352-.251h-3.593z"/>
-    </svg>
-
-    <svg v-else-if="isPing" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 64 48" style="transform: scaleY(-1)">
-      <path d="M20 26a2 2 0 0 0 1.93-1.474l4.077-14.9 10.063 36.9c.525 1.923 3.232 1.98 3.836.08L46.45 26H62a2 2 0 0 0 0-4H45a2 2 0 0 0-1.906 1.394l-4.942 15.56L27.93 1.475c-.536-1.965-3.324-1.965-3.86 0L18.5 22H2a2 2 0 0 0 0 4h18z"
-            stroke="currentColor" stroke-width="0.3"></path>
-    </svg>
-
-    <svg v-else-if="isPort" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 30 30" style="transform: scaleY(-1)">
-      <g style="transform: translate(-3px, -3px)">
-        <path d="M19 27h-8c-.6 0-1-.4-1-1v-2H7.9c-.6 0-1-.4-1-1v-2.7H4.3c-.6 0-1-.4-1-1V4c0-.6.4-1 1-1h21.4c.6 0 1 .4 1 1v15.4c0 .6-.4 1-1 1h-2.6v2.7c0 .6-.4 1-1 1H20v2c0 .5-.5.9-1 .9zm-7-2h6v-2c0-.6.4-1 1-1h2.1v-2.7c0-.6.4-1 1-1h2.6V5H5.3v13.4h2.6c.6 0 1 .4 1 1v2.7H11c.6 0 1 .4 1 1V25z"></path>
-        <path d="M21.9 12.3c-.6 0-1-.4-1-1V7.9c0-.6.4-1 1-1s1 .4 1 1v3.4c0 .6-.4 1-1 1zM17.3 12.3c-.6 0-1-.4-1-1V7.9c0-.6.4-1 1-1s1 .4 1 1v3.4c0 .6-.4 1-1 1zM12.7 12.3c-.6 0-1-.4-1-1V7.9c0-.6.4-1 1-1s1 .4 1 1v3.4c0 .6-.4 1-1 1zM8.1 12.3c-.6 0-1-.4-1-1V7.9c0-.6.4-1 1-1s1 .4 1 1v3.4c0 .6-.4 1-1 1z"></path>
-      </g>
-    </svg>
-
-    <svg v-else-if="isPingdom" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 128 128">
-      <path d="M60.938 60.755l1.102-.308a212.455 212.455 0 0 0 24.113-8.85 89.068 89.068 0 0 0 22.192-12.646A48.251 48.251 0 0 0 123.9 16.942a19.734 19.734 0 0 0 1.409-4.463c-10.788 14.749-53.557 16.289-53.557 16.289l10.096-9.261A140.85 140.85 0 0 0 0 46.082a77.044 77.044 0 0 1 43.41 19.419c5.791-1.796 11.711-3.079 17.528-4.695v-.051zM128 59.293a235.356 235.356 0 0 0-80.464 10.261 77.582 77.582 0 0 1 13.53 19.315C82.617 76.864 112.496 61.525 128 59.344v-.051zM63.782 95a77.006 77.006 0 0 1 4.869 20.521l37.593-38.657L63.782 95z"/>
-    </svg>
-
-    <svg v-else-if="isGitHub" viewBox="0 0 33 33" xmlns="http://www.w3.org/2000/svg" fill-rule="evenodd" clip-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" style="transform: translate(-1px, 2px)">
-      <path d="M16.288 0C7.293 0 0 7.293 0 16.29c0 7.197 4.667 13.302 11.14 15.457.815.149 1.112-.354 1.112-.786 0-.386-.014-1.411-.022-2.77-4.531.984-5.487-2.184-5.487-2.184-.741-1.881-1.809-2.382-1.809-2.382-1.479-1.011.112-.991.112-.991 1.635.116 2.495 1.679 2.495 1.679 1.453 2.489 3.813 1.77 4.741 1.354.148-1.053.568-1.771 1.034-2.178-3.617-.411-7.42-1.809-7.42-8.051 0-1.778.635-3.232 1.677-4.371-.168-.412-.727-2.068.159-4.311 0 0 1.368-.438 4.48 1.67a15.602 15.602 0 014.078-.548c1.383.006 2.777.187 4.078.548 3.11-2.108 4.475-1.67 4.475-1.67.889 2.243.33 3.899.162 4.311 1.044 1.139 1.675 2.593 1.675 4.371 0 6.258-3.809 7.635-7.438 8.038.585.503 1.106 1.497 1.106 3.017 0 2.177-.02 3.934-.02 4.468 0 .436.293.943 1.12.784 6.468-2.159 11.131-8.26 11.131-15.455C32.579 7.293 25.285 0 16.288 0" fill="currentColor"/>
-    </svg>
-
-    <svg v-else-if="isGitLab" viewBox="0 0 56 52" xmlns="http://www.w3.org/2000/svg">
-      <path d="M52.71 30.572L35.68 43.007l4.566-5.885 10.561-13.605 2.091 6.459a.53.53 0 0 1-.187.596m-49.49 0a.532.532 0 0 1-.19-.595l2.093-6.461 15.13 19.497-17.03-12.441M11.395 4.1l4.772 14.745H6.639L11.395 4.1m24.26 17.745l-7.682 23.769-7.684-23.769h15.366m8.887-17.744l4.765 14.744h-9.53l4.765-14.744M32.029 42.82l6.777-20.964H48.3L32.471 42.251l-.442.569M17.132 21.856l6.775 20.959L7.646 21.856h9.486m38.623 7.197l-2.944-9.09c-.006-.022-.006-.044-.013-.066l-5.872-18.17C46.58.692 45.579-.031 44.525.007a2.505 2.505 0 0 0-2.373 1.736l-5.527 17.11H19.318L13.78 1.743A2.502 2.502 0 0 0 11.409.008c-1.113-.041-2.054.686-2.407 1.737l-5.85 18.14-.007.015-2.967 9.16a3.53 3.53 0 0 0 1.273 3.941L27.075 51.72c.054.039.115.056.171.087l.016.01.01.007a1.464 1.464 0 0 0 .342.13l.002.001.05.009c.027.006.055.007.082.011.07.01.14.033.211.033.158 0 .315-.032.467-.081l.011-.003.009-.003c.055-.019.109-.041.162-.066.017-.008.033-.019.05-.027l.003-.002.005-.002.003-.002a1.6 1.6 0 0 0 .168-.102l.02-.01.002-.001L54.484 33a3.528 3.528 0 0 0 1.271-3.941"></path>
-    </svg>
-
-    <svg v-else-if="isTravis" viewBox="0 0 200 199" xmlns="http://www.w3.org/2000/svg" stroke="currentColor" stroke-width="1.2" stroke-linejoin="round">
-      <g style="transform: translateX(-14px)">
-        <path d="M138.347 116.424a2.203 2.203 0 0 0 2.154-2.235 2.186 2.186 0 0 0-2.235-2.153 2.194 2.194 0 0 0 .081 4.388zm13.488 9.878c-2.89-1.396-6.059.377-11.828.484-4.292.151-7.896.198-8.132-6.543-.237-6.747 2.513-12.326 6.805-12.478 4.292-.15 8.207 5.2 8.735 11.931.145 1.854-.06 3.207-.521 4.21 3.996-.477 4.899 2.235 4.941 2.396z"></path>
-        <path d="M76.85 118.051a2.195 2.195 0 1 0-.122-4.388 2.195 2.195 0 0 0 .122 4.388zm6.634 3.529c.063 6.747-3.509 6.339-7.806 6.381-.435.011-.848.016-1.258.022-.482.011-.944.016-1.39.005-4.168-.005-6.833-.194-9.19 1.079.058-.145 1.09-2.461 4.835-3.4-.414-.914-.673-2.181-.742-3.937-.257-6.741 3.9-12.269 8.197-12.306 4.292-.042 7.289 5.411 7.354 12.156zM73.568 72.27a1.65 1.65 0 0 1-.098-.247c-1.975-7.187-1.889-14.723-1.712-18.199l-20.514 5.833c-1.38 8.982-.736 15.583-.597 16.797 1.703-.462 3.411-.887 5.131-1.284 2.836-.628 5.693-1.154 8.556-1.638 2.869-.478 5.747-.843 8.626-1.192.205-.027.404-.042.608-.07zm8.933-.864c-2.408-5.05-2.934-10.422-3.046-12.317-1.858-1.316-3.696-2.895-5.049-4.125-.119 3.631-.032 10.206 1.654 16.345.065.237.058.473 0 .694 2.145-.227 4.292-.425 6.441-.597zm56.78-.098c1.688-6.15 1.779-12.72 1.655-16.345-1.353 1.23-3.197 2.809-5.049 4.125-.114 1.896-.64 7.278-3.052 12.332 2.149.173 4.298.366 6.446.591a1.33 1.33 0 0 1 0-.703zm25.409 5.151c.14-1.207.789-7.809-.591-16.801l-20.52-5.833c.184 3.475.265 11.012-1.707 18.199a1.619 1.619 0 0 1-.101.258c.203.021.408.037.606.064 5.769.661 11.511 1.584 17.189 2.83 1.712.398 3.426.823 5.124 1.283zm3.297.968c2.755.902 5.526 1.858 8.036 3.325-1.343-.532-2.729-.913-4.126-1.257a70.385 70.385 0 0 0-4.201-.924c-2.82-.531-5.65-.982-8.498-1.327-2.841-.37-5.687-.682-8.546-.924-2.858-.241-5.709-.483-8.573-.65-11.447-.704-22.924-.88-34.41-.892-11.483.006-22.962.221-34.409.897-2.862.166-5.715.409-8.572.651-2.857.241-5.71.548-8.546.923-2.847.345-5.678.796-8.498 1.327-1.407.264-2.81.57-4.206.919-1.391.344-2.783.725-4.126 1.257 2.509-1.466 5.28-2.427 8.041-3.331.232-.075.467-.139.703-.214-.015-.059-.032-.113-.043-.177-.048-.317-1.069-7.859.709-18.645.086-.516.456-.935.962-1.075l2.917-.831c.634-22.625 9.952-33.266 10.243-33.594-8.326 13.397-8.25 29.286-8.106 32.986l18.128-5.152c.016-.005.026-.005.042-.01.076-.016.151-.027.226-.032.021 0 .049-.006.075-.006a1.19 1.19 0 0 1 .297.027c.015 0 .031.011.053.016.075.016.145.042.224.075.033.016.054.033.086.049.058.033.119.07.177.112.016.011.034.016.049.033l.032.032c.016.016.037.027.054.044.012.016.494.493 1.262 1.209-.182-5.973.102-23.108 8.262-37.31-.172.498-6.646 19.428-4.415 40.645.724.58 1.486 1.149 2.229 1.649.359.247.58.655.585 1.09.006.07.161 6.833 3.148 12.586.042.086.074.177.102.268 7.429-.505 14.878-.709 22.312-.714 7.436.005 14.88.22 22.307.731.027-.097.06-.193.109-.285 2.986-5.753 3.142-12.516 3.142-12.586.01-.436.231-.843.591-1.09.741-.5 1.493-1.069 2.224-1.649 2.234-21.217-4.24-40.147-4.411-40.645 8.153 14.201 8.444 31.336 8.262 37.31a62.536 62.536 0 0 0 1.261-1.209c.016-.016.039-.027.053-.044.012-.01.018-.021.033-.032.016-.016.033-.022.049-.033.06-.042.119-.079.177-.118.028-.01.054-.027.081-.043.081-.033.155-.059.236-.08.016 0 .033-.011.049-.011.096-.021.2-.032.296-.027.027 0 .049.006.07.006.075.005.156.016.231.032.012.006.028.006.042.01l18.129 5.152c.146-3.7.221-19.59-8.104-32.986.289.328 9.609 10.969 10.237 33.594l2.922.831c.499.14.875.559.962 1.075 1.777 10.786.752 18.328.708 18.645-.01.065-.026.124-.042.182.239.07.47.139.707.215z"></path>
-        <path d="M129.295 19.597v17.025h-12.069v-4.973h-2.719v22.98h4.108V67.48H97.506V54.629h4.092v-22.98h-2.711v4.973h-12.06V19.597h42.468z"></path>
-        <path d="M37.745 92.276a229.005 229.005 0 0 0-7.051 1.47l.829 3.127a103.93 103.93 0 0 1 6.222-4.597zm3.815 51.36a68.861 68.861 0 0 1-.621-2.102l-5.162-3.612a1.391 1.391 0 0 1-.586-.969l-2.516-20.449c-1.864.999-4.017 2.225-4.592 2.707-.497.409-.875 1.46-.336 6.332.668 6.01 3.712 17.333 5.348 18.979 1.968.347 6.327-.258 8.465-.886zm30.85-54.197a150.122 150.122 0 0 1 11.813-2.796c-5.798.212-11.6.481-17.393.808-3.366.186-6.715.414-10.065.667-1.678.129-3.345.263-5.007.445-.476.046-.942.098-1.418.16-4.369 2.614-21.127 13.134-32.631 26.889 11.179-7.769 30.654-19.443 54.701-26.173zm-2.159 9.949c3.251-3.225 6.605-5.886 9.567-7.967-11.415 2.651-21.923 6.543-31.128 10.778a360.846 360.846 0 0 1 21.561-2.811zm43.814-13.144c-2.31-.036-4.623-.057-6.933-.062h-.005c-3.39.005-6.787.041-10.189.109l-6.269 2.971c-.005.005-.041.021-.088.048-.942.46-9.174 4.613-16.919 12.021 6.943-3.65 17.146-8.418 29.153-12.115a144.186 144.186 0 0 1 11.25-2.972zm-3.816 18.788c1.827-1.103 3.582-2.366 5.249-3.712a422.33 422.33 0 0 0-7.278-.072c-10.137 0-19.606.415-28.189 1.061-8.61 4.209-13.875 7.672-13.998 7.76l-8.268 5.514 5.679-8.149a52.176 52.176 0 0 1 2.956-3.857c-9.536 1.066-17.477 2.329-23.41 3.422l3.038 24.632 10.453 7.321 33.184-7.377 6.212-20.465c.104-.337.331-.621.627-.793.098-.063.201-.109.315-.14.192-.052 3.51-.999 7.336-1.465l6.094-3.68zm3.354 78.632a45.604 45.604 0 0 0 3.184-.378 138.958 138.958 0 0 1-3.568-3.857 398.441 398.441 0 0 1-1.92 4.339h.243c.658.001 1.378-.071 2.061-.104zm17.4 7.461c-4.503-.332-8.598-2.205-11.903-5.477a271.86 271.86 0 0 0-.502-.512c-1.598.321-3.22.559-4.881.704-.698.026-1.361.087-2.091.087l-1.083.011-.413-.011c-4.396 6.539-14.159 7.813-14.605 7.87-.403.046-.734.103-1.191.186 5.442 1.491 10.996 2.138 16.474 1.77 5.492-.367 12.627-1.558 20.195-4.628zM127.416 89.04c.377-.471.729-.926 1.044-1.34-3.281.331-6.512.808-9.67 1.408-10.814 2.024-20.801 5.389-29.11 8.837a383.259 383.259 0 0 1 18.54-.455c3.908 0 7.708.067 11.404.176 3.179-3.056 5.861-6.182 7.792-8.626zm29.876-.926c-.01-.006-.02-.012-.031-.012a551.624 551.624 0 0 0-9.826-.651 905.6 905.6 0 0 0-13.667-.668 72.95 72.95 0 0 1-1.574 2.225c-2.479 3.355-7.398 9.51-13.704 14.729 8.926-1.6 24.409-5.56 37.803-14.905.336-.238.668-.486.999-.718zm15.034 51.593l-1.404.983-3.779 2.651-4.095 2.868c-.279.192-.621.28-.954.243l-40.746-4.633-2.966-.337a1.39 1.39 0 0 1-1.171-.977l-6.377-20.998c-1.066-.145-2.014-.219-2.81-.219-.809 0-1.751.073-2.817.219l-6.192 20.392a1.383 1.383 0 0 1-1.025.946l-43.435 9.672c-.103.02-.206.03-.305.03-.279 0-.559-.083-.798-.253l-1.578-1.098-4.726-3.307v-.011l-1.91-1.335c.135.43.289.85.441 1.268l-.006.006c.368 1.092 4.028 11.622 11.467 21.929a873.96 873.96 0 0 0 17.057-5.234c4.488-6.917 10.877-13.777 15.418-14.014a51.12 51.12 0 0 1 2.402-.061c2.221 0 4.344.16 6.31.393-1.671-1.517-2.013-3.298-2.256-4.085 0 0 5.793 4.53 13.17 3.584 7.378-.953 11.959-5.204 11.959-5.204s-.021 3.102-3.236 5.503c6.182-1.869 8.739-2.511 10.489-2.511 1.931 0 2.883.808 3.717 1.519.161.129.321.268.507.419a3519.302 3519.302 0 0 1 12.141 9.614c1.936.227 8.075.926 11.943 1.283 2.231.201 4.245.372 6.217.372.637 0 1.233-.026 1.797-.063 2.558-4.88 4.857-10.411 6.808-16.653.261-.96.516-1.928.743-2.901zm-2.303-11.845l2.992-20.948c-8.439-1.536-20.78-3.394-35.897-4.554-13.647 4.707-25.077 6.108-25.766 6.155l-.797.057c4.353.374 8.454 1.544 8.66 1.605.452.135.804.481.944.933l6.186 20.366 33.138 3.764 10.54-7.378zm8.07 4.411l3.385-23.676a240.127 240.127 0 0 0-5.731-1.169l-3.059 21.423a1.415 1.415 0 0 1-.575.942l-11.472 8.023c-.27.192-.616.28-.947.243l-34.572-3.929a1.391 1.391 0 0 1-1.176-.973l-6.227-20.5c-1.668-.431-5.949-1.43-9.696-1.43-3.764 0-8.041.999-9.708 1.43l-6.228 20.5a1.388 1.388 0 0 1-1.025.947l-34.572 7.692a1.483 1.483 0 0 1-.306.033 1.36 1.36 0 0 1-.792-.25l-11.467-8.029a1.396 1.396 0 0 1-.585-.968l-3.091-25.072c-1.284.249-2.443.487-3.479.703-.734.405-1.46.809-2.174 1.213l3.281 26.568 16.666 11.675 42.047-9.354 6.207-20.449a1.389 1.389 0 0 1 1.108-.975c1.574-.253 2.95-.382 4.116-.382 1.153 0 2.536.129 4.105.382.528.083.957.461 1.108.975l6.366 20.956 42.282 4.808 16.211-11.357zm4.68-27.237l2.977-11.29c-4.318-.978-12.27-2.615-23.1-4.148-5.53 3.976-11.582 7.155-17.53 9.691 18.199 1.771 31.57 4.406 37.653 5.747zm5.161 17.648c.456-5.052.058-6.139-.455-6.554-.513-.43-2.247-1.412-3.935-2.329l-2.779 19.464a1.39 1.39 0 0 1-.58.942l-3.977 2.781c-.315 1.593-.429 2.345-.817 3.903 2.273.528 5.999.938 7.775.595 1.612-1.748 4.214-12.61 4.768-18.802zm11.995-48.031a53.972 53.972 0 0 1-.258 8.733 55.462 55.462 0 0 1-1.619 8.605c-.4 1.414-.86 2.811-1.404 4.198a38.295 38.295 0 0 1-.89 2.071c-.161.341-.331.678-.523 1.025l-.284.512a8.975 8.975 0 0 1-.348.574l-.294.457-.461.237c-.492.254-.895.445-1.342.653l-1.298.585a88.22 88.22 0 0 1-2.62 1.065c-.611.239-1.15.457-1.662.674l-1.444 5.487c-.036-.009-.471-.12-1.283-.315l-.078.574c1.594.833 4.726 2.522 5.793 3.403 2.148 1.775 2.299 4.587 1.823 9.841-.244 2.697-1.139 7.946-2.381 12.767-2.144 8.298-3.283 9.273-4.753 9.649-.746.192-1.894.383-3.008.383-2.266 0-5.353.063-7.429-.439-.533 1.888-2.055 6.812-5.068 12.962.151-.073.3-.135.435-.207 3.717-1.952 10.861-5.064 11.162-5.199l5.643-2.452-2.89 5.435c-.067.118-6.264 11.773-10.059 19.383-3.769 7.538-10.835 9.179-15.065 10.151-.637.151-1.241.291-1.733.425-3.035.854-9.18 2.319-14.599 3.623-.064.016-.13.033-.197.042a64.057 64.057 0 0 1-10.955 5.411c-14.568 5.518-29.923 5.208-43.844.092a647.05 647.05 0 0 1-9.193 1.097 45.12 45.12 0 0 1-4.985.291c-13.264 0-20.294-6.736-25.425-13.331-5.493-7.062-12.212-17.546-12.497-17.985L31 159.053l6.585 2.961c3.152 1.419 12.524 5.757 15.205 7 .217-.061.43-.124.642-.186-4.457-6.357-8.112-13.605-10.695-21.634-2.195.662-5.576 1.175-8.206 1.175-.961 0-1.822-.072-2.484-.228-1.471-.336-3.148-1.754-5.431-9.795-1.325-4.668-2.314-9.764-2.603-12.387-.57-5.121-.466-7.864 1.662-9.636 1.283-1.071 5.611-3.344 6.507-3.809l-.192-1.58c-13.75 8.08-21.991 15.22-22.157 15.366L0 134.929l7.005-11.047c5.544-8.755 11.948-15.832 17.84-21.284-.244-.098-.471-.196-.71-.294l-1.299-.585a34.907 34.907 0 0 1-1.34-.653l-.461-.237-.295-.457c-.166-.249-.238-.388-.347-.574l-.29-.512c-.181-.347-.358-.684-.518-1.025a30.878 30.878 0 0 1-.89-2.071 44.74 44.74 0 0 1-1.404-4.198 54.745 54.745 0 0 1-1.62-8.605 54.664 54.664 0 0 1-.259-8.733c.078-1.455.218-2.909.419-4.354.103-.725.213-1.45.357-2.17.151-.734.297-1.418.519-2.221l.155-.564.404-.317c2.294-1.802 4.768-3.163 7.284-4.369a78.87 78.87 0 0 1 6.311-2.616c5.943-16.493 16.162-31.118 29.591-41.311C74.337 6.197 90.664.627 107.671.627s33.334 5.57 47.218 16.106c13.43 10.193 23.649 24.819 29.588 41.307a78.282 78.282 0 0 1 6.316 2.62c2.515 1.206 4.99 2.567 7.283 4.369l.404.317.156.564c.227.803.372 1.487.517 2.221.146.72.26 1.445.357 2.17.203 1.443.348 2.897.419 4.352zM40.499 56.337c-2.516 1.014-5.016 2.06-7.46 3.209-2.449 1.119-4.856 2.32-7.155 3.66-2.121 1.222-4.157 2.563-5.954 4.076-.077.455-.149.952-.211 1.423a51.357 51.357 0 0 0-.388 6.068c-.026 2.713.16 5.426.502 8.112.372 2.692.864 5.369 1.594 7.952a41.963 41.963 0 0 0 1.243 3.804c.233.623.492 1.228.762 1.818.134.294.274.585.413.864l.172.326c.201.104.409.207.605.3l1.206.574c.673.311 1.6.751 2.366 1.093.046-.037.088-.078.124-.114l-2.231-8.511c.471-.129 4.717-1.227 12.032-2.619a33.744 33.744 0 0 1-1.775-.379 36.704 36.704 0 0 1-4.898-1.563 22.857 22.857 0 0 1-2.309-1.119c-.741-.425-1.471-.905-2.035-1.547 8.035 2.624 24.637 1.433 39.398-.088 13.501-1.393 27.028-2.293 40.628-2.325 13.6.031 27.138.931 40.63 2.325 14.77 1.522 31.374 2.713 39.406.088-.564.642-1.293 1.122-2.034 1.547-.739.42-1.522.782-2.309 1.119a36.965 36.965 0 0 1-4.903 1.563c-.244.056-.492.114-.741.166 8.02 1.486 12.689 2.697 13.186 2.832l-2.138 8.107c.43-.192.864-.377 1.288-.574l1.207-.574c.196-.094.404-.196.606-.3l.166-.326c.144-.279.284-.57.419-.864.27-.591.528-1.196.767-1.818.471-1.231.879-2.51 1.236-3.804.731-2.583 1.228-5.26 1.595-7.952.346-2.686.528-5.4.502-8.112a52.755 52.755 0 0 0-.176-4.059 51.573 51.573 0 0 0-.213-2.009 29.83 29.83 0 0 0-.213-1.423c-1.797-1.513-3.831-2.853-5.954-4.076-2.299-1.34-4.704-2.541-7.159-3.66-2.438-1.149-4.943-2.195-7.46-3.209a138.835 138.835 0 0 0-3.801-1.476c-1.267-.491-2.552-.956-3.835-1.423 2.696.445 5.369 1.06 8.013 1.739 1.724.446 3.444.948 5.141 1.481-12.11-31.658-41.07-52.272-72.685-52.272-31.622 0-60.576 20.614-72.684 52.272a107.832 107.832 0 0 1 5.135-1.481c2.651-.678 5.322-1.294 8.02-1.739-1.29.466-2.568.931-3.842 1.423-1.268.47-2.535.967-3.799 1.475zm72.127 115.799l1.594 1.899c.036.046 3.577 4.26 7.906 8.552 2.879 2.853 6.357 4.297 10.343 4.297 1.361 0 2.791-.175 4.235-.523 1.34-.326 2.796-.673 4.287-1.03 5.384-1.287 11.482-2.749 14.438-3.577.585-.166 1.238-.315 1.925-.472 3.935-.909 9.329-2.163 12.187-7.889 2.149-4.297 5.047-9.874 7.197-13.961-1.863.859-3.816 1.79-5.203 2.52-2.138 1.123-4.938 1.667-8.558 1.667-2.152 0-4.266-.181-6.605-.389-4.675-.43-12.586-1.361-12.667-1.372l-.606-.067-.478-.383c-.071-.052-7.003-5.575-12.606-9.981a75.727 75.727 0 0 1-.621-.513c-.59-.503-.59-.503-.943-.503-1.796 0-7.019 1.62-18.461 5.167l-.703.223-.689-.26c-.078-.026-7.585-2.81-16.581-2.81-.736 0-1.47.019-2.185.056-.901.046-5.958 2.448-12.425 12.68l-.419.657-.741.238c-.107.037-11.238 3.63-23.042 7.005l-.766.218-.725-.337c-.077-.031-4.696-2.174-9.091-4.194 2.397 3.541 5.462 7.958 8.159 11.422 4.711 6.067 10.649 11.674 22.034 11.674 1.428 0 2.945-.088 4.503-.265 11.581-1.309 14.563-1.837 16.168-2.117.543-.092.973-.171 1.522-.238.088-.011 9.571-1.237 12.232-7.206 2.744-6.134 3.298-7.595 3.319-7.651l.968-2.583-.263-.245c.192-.207.383-.424.58-.632 0 .005 0 .005.005.005l.019.016c.306.218.611.45.916.684l-.159.218z"></path>
-      </g>
-    </svg>
-
-    <svg v-else-if="isJenkins" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 128 128">
-      <path d="M127.15 104.133l-19.926-14.012c7.85-9.503 12.775-23.071 12.775-38.119 0-28.671-17.863-51.997-39.82-51.997a30.814 30.814 0 0 0-6.404.681 1.98 1.98 0 0 0-.518-.15C50.843-2.42 40.509 7.686 38.21 10.353c-9.702 2.47-12.73 10.147-13.079 14.684-.441.123-.81.23-1.08.31-12.501 4.261-6.903 24.066-6.288 26.11a20.664 20.664 0 0 0-2.772 10.545C14.99 71.375 20.72 79 27.762 79c.234 0 .465-.017.697-.033a51.953 51.953 0 0 0 5.413 12L1.409 111.314a3.057 3.057 0 0 0-1.011 4.094l6.895 11.613a2 2 0 1 0 3.44-2.042l-6.4-10.777 9.702-6.08 13.318 19.025a2 2 0 0 0 3.278-2.294l-13.202-18.858L26 100.622a41.22 41.22 0 0 0 34.653 18.533q.564 0 1.14-.013a1.963 1.963 0 0 0 .214-.027l-.014 4.918a3.917 3.917 0 0 0 1.873 3.357 4.014 4.014 0 0 0 3.93.182l10.534-5.252a7.496 7.496 0 0 0 4.561 1.875c.145.008.288.012.431.012a7.473 7.473 0 0 0 4.962-1.87c.168-.148.326-.304.479-.465l11.43 5.7a4.012 4.012 0 0 0 3.932-.182 3.917 3.917 0 0 0 1.872-3.357l-.017-6.147a22.458 22.458 0 0 0 13.411-14.318l4.186 2.944-7.127 18.777a2 2 0 1 0 3.74 1.42l7.679-20.23a1.999 1.999 0 0 0-.72-2.346zm-87.383-90.04a2.024 2.024 0 0 0 1.177-.766c.315-.423 7.037-9.164 23.938-9.326-7.758 4.231-14.278 11.578-18.64 20.838-4.082-2.762-11.631-1.907-16.909-.816.61-2.924 2.771-8.23 10.434-9.93zm21.929 101.05q-.52.013-1.032.012A37.216 37.216 0 0 1 29.39 98.498l6.73-4.217c5.164 6.96 13.285 13.664 25.706 14.712.057.004.115.006.17.006.014 0 .027-.003.04-.003l-.017 6.165a1.794 1.794 0 0 0-.323-.018zm2.239-10.55a4.004 4.004 0 0 0-.872.726 1.98 1.98 0 0 0-.9-.313c-21.175-1.785-28.304-21.652-29.866-27.095a11.856 11.856 0 0 0 2.458-1.653 2 2 0 1 0-2.62-3.023A6.687 6.687 0 0 1 27.763 75c-4.754 0-8.771-5.953-8.771-13s4.017-12.999 8.771-12.999a7.365 7.365 0 0 1 5.663 3.123 2 2 0 0 0 3.135-2.485 11.261 11.261 0 0 0-8.798-4.638 10.527 10.527 0 0 0-6.817 2.65c-1.109-5.199-2.545-16.153 4.326-18.495 6.478-1.936 16.724-3.397 19.229-.586.034.038.074.066.11.1a64.857 64.857 0 0 0-4.25 23.331 66.562 66.562 0 0 0 2.105 16.75h-4.473a2 2 0 1 0 0 4h6.967a2 2 0 0 0 2-2.037l-.01-.53a2.032 2.032 0 0 0-.082-.531 62.16 62.16 0 0 1-2.507-17.652c0-26.466 16.069-47.998 35.82-47.998 12.221 0 23.028 8.25 29.495 20.807a12.877 12.877 0 0 0-10.72.812 2 2 0 0 0 2.085 3.413c.234-.144 5.762-3.41 11.466 2.309A61.514 61.514 0 0 1 116 52.002c0 26.466-16.069 47.998-35.82 47.998a29.425 29.425 0 0 1-19.099-7.41 2 2 0 0 0-2.588 3.051A33.374 33.374 0 0 0 80.18 104c9.163 0 17.606-4.073 24.342-10.89l4.598 3.234c-.017.049-.04.094-.054.145a17.45 17.45 0 0 1-5.327 7.936 4 4 0 0 0-3.617.015l-12.489 6.35a7.517 7.517 0 0 0-8.044-.392l-11.717-5.958a4.01 4.01 0 0 0-3.936.153zm2.06 19.451l.063-16.039 10.61 5.396a7.204 7.204 0 0 0-.514 5.55zm20.785-7.039a3.328 3.328 0 0 1-1.14 2.33 3.5 3.5 0 0 1-2.525.866 3.394 3.394 0 0 1 .177-6.785c.066 0 .133.002.2.006a3.5 3.5 0 0 1 2.414 1.142 3.325 3.325 0 0 1 .874 2.44zm3.841 1.324a7.406 7.406 0 0 0 .154-1.102 7.283 7.283 0 0 0-.576-3.266l11.754-6.007.027 16.038zm15.348-4.778l-.016-5.61c0-.044-.007-.087-.009-.13a22.179 22.179 0 0 0 6.579-9.074l3.41 2.399a18.496 18.496 0 0 1-9.964 12.415z"></path>
-      <path d="M79.996 69.219a6.275 6.275 0 0 0 2.69 5.353 10.044 10.044 0 0 0 5.665 1.422 31.612 31.612 0 0 0 9.656-1.87c4.814-1.526 7.481-3.773 7.927-6.68.678-4.42-4.201-7.982-4.76-8.374a2.079 2.079 0 0 0-.204-.126c-7.402-3.963-9.935-17.172-9.96-17.306a2 2 0 0 0-3.934.727c.114.617 2.88 15.105 11.888 20.041 1.103.804 3.25 2.907 3.016 4.432-.188 1.223-2.076 2.489-5.266 3.502-3.307 1.21-9.431 2.534-11.772.93a2.31 2.31 0 0 1-.946-2.128 2 2 0 0 0-4 .077zM61.196 32.101c1.884-1.221 4.325-1.422 7.257-.595a2 2 0 0 0 1.084-3.85c-4.087-1.15-7.626-.784-10.516 1.089-4.96 3.215-5.966 9.692-6.006 9.967a2 2 0 1 0 3.958.582c.007-.05.819-4.987 4.223-7.193zM69.995 46.148a2 2 0 0 0-2.604 1.104l-1.25 3.088a2 2 0 0 0 1.324 2.678 24.595 24.595 0 0 0 6.498.977 12.123 12.123 0 0 0 5.624-1.197 4.354 4.354 0 0 0 2.254-2.595 3.765 3.765 0 0 0-.356-2.96l-4.694-8.231a2 2 0 0 0-3.474 1.981l4.63 8.122a.992.992 0 0 1-.196.128c-.987.51-3.124 1.154-7.025.43l.373-.92a2 2 0 0 0-1.104-2.605zM99.305 51.655a19.348 19.348 0 0 0 6.673 1.344 10.006 10.006 0 0 0 5.832-1.67 5.1 5.1 0 0 0 1.781-6.164 2.009 2.009 0 0 0-.16-.304l-5.78-8.944a2 2 0 0 0-3.36 2.171l5.662 8.763a1.07 1.07 0 0 1-.42 1.19c-1.157.802-3.27 1.456-6.952.468l.298-.829a2 2 0 0 0-3.764-1.355l-1 2.776a2 2 0 0 0 1.19 2.554zM88.668 78.8c-8.383 1.391-13.92-4.833-14.16-5.107a2 2 0 0 0-3.03 2.612c.261.302 5.876 6.699 14.743 6.699a18.869 18.869 0 0 0 3.104-.26 2 2 0 1 0-.657-3.945zM105.485 74.692c-1.219 1.409-4.509 2.655-8.8 3.333a2 2 0 0 0 .31 3.976 2.03 2.03 0 0 0 .315-.025c3.785-.599 8.801-1.894 11.2-4.666a2 2 0 0 0-3.025-2.618zM88.688 88.596c-14.958 2.334-19.716-6.142-19.923-6.525a2 2 0 0 0-3.547 1.848c.214.415 4.872 9.086 18.453 9.086a36.565 36.565 0 0 0 5.634-.457 2 2 0 1 0-.617-3.952zM23.902 56.318a2 2 0 0 0 1.162 3.828 3.45 3.45 0 0 1 4.525 2.474 2 2 0 0 0 3.805-1.237 7.464 7.464 0 0 0-9.491-5.065z"></path>
-    </svg>
-
-    <svg v-else-if="isAzure" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 256 256">
-      <path d="M193.412 58.029l62.586-11.421v157.685l.005.001-62.588 51.604-.004-.001-.001.001v-.001l-96.473-32.233V256L25.94 172.124l167.47 23.417V58.03L25.938 89.904v82.22L0 167.306V94.841l25.938-33.88 87.979-34.394V0l79.495 58.029z" fill-rule="nonzero"/>
+    <svg viewBox="0 0 64 64">
+      <use :xlink:href="'./icons.svg#' + iconId"></use>
     </svg>
   </div>
 </template>
 
 <script lang="ts">
-  import Vue from 'vue'
-  import {Component, Prop} from 'vue-property-decorator'
+import {computed, defineComponent} from 'vue'
 
-  import TileType from '@/enums/tileType'
+import TileIconId from '@/enums/tileIconId'
+import TileType from '@/enums/tileType'
 
-  @Component
-  export default class MonitororTileIcon extends Vue {
-    /*
-     * Props
-     */
+export default defineComponent({
+  name: 'MonitororTileIcon',
+  props: {
+    tileType: {
+      type: String as () => TileType,
+      required: true,
+    },
+  },
+  setup(props) {
+    const iconId = computed(() => {
+      switch (props.tileType) {
+        case TileType.AzureDevOpsBuild:
+        case TileType.AzureDevOpsRelease:
+          return TileIconId.Azure
 
-    @Prop()
-    private tileType!: TileType
+        case TileType.GitHubChecks:
+        case TileType.GitHubPullRequest:
+        case TileType.GitHubCount:
+          return TileIconId.GitHub
 
-    /*
-     * Computed
-     */
+        case TileType.GitLabIssues:
+        case TileType.GitLabMergeRequest:
+        case TileType.GitLabPipeline:
+          return TileIconId.GitLab
 
-    get isGroup(): boolean {
-      return this.tileType === TileType.Group
-    }
+        case TileType.Group:
+          return TileIconId.Group
 
-    get isHttp(): boolean {
-      return [
-        TileType.HttpStatus,
-        TileType.HttpRaw,
-        TileType.HttpFormatted,
-      ].includes(this.tileType)
-    }
+        case TileType.JenkinsBuild:
+          return TileIconId.Jenkins
 
-    get isPing(): boolean {
-      return this.tileType === TileType.Ping
-    }
+        case TileType.HttpStatus:
+        case TileType.HttpRaw:
+        case TileType.HttpFormatted:
+          return TileIconId.Http
 
-    get isPort(): boolean {
-      return this.tileType === TileType.Port
-    }
+        case TileType.Ping:
+          return TileIconId.Ping
 
-    get isPingdom(): boolean {
-      return [
-        TileType.PingdomCheck,
-        TileType.PingdomTransactionCheck,
-      ].includes(this.tileType)
-    }
+        case TileType.PingdomCheck:
+        case TileType.PingdomTransactionCheck:
+          return TileIconId.Pingdom
 
-    get isGitHub(): boolean {
-      return [
-        TileType.GitHubChecks,
-        TileType.GitHubPullRequest,
-        TileType.GitHubCount,
-      ].includes(this.tileType)
-    }
+        case TileType.Port:
+          return TileIconId.Port
 
-    get isGitLab(): boolean {
-      return [
-        TileType.GitLabPipeline,
-        TileType.GitLabMergeRequest,
-        TileType.GitLabIssues,
-      ].includes(this.tileType)
-    }
+        case TileType.TravisCiBuild:
+          return TileIconId.TravisCi
+      }
+    })
 
-    get isTravis(): boolean {
-      return this.tileType === TileType.Travis
-    }
-
-    get isJenkins(): boolean {
-      return this.tileType === TileType.Jenkins
-    }
-
-    get isAzure(): boolean {
-      return [
-        TileType.AzureDevOpsBuild,
-        TileType.AzureDevOpsRelease,
-      ].includes(this.tileType)
+    return {
+      iconId,
     }
   }
+})
 </script>
 
 <style lang="scss">
