@@ -1,13 +1,15 @@
 import TileStatus from '@/enums/tileStatus'
 import TileType from '@/enums/tileType'
-import TileBuild from '@/interfaces/tileBuild'
-import TileValue from '@/interfaces/tileValue'
+import TileBuild from '@/types/tileBuild'
+import TileMetrics from '@/types/tileMetrics'
 
-export default interface TileState {
+type TileState = {
   type: TileType,
   status: TileStatus,
   label?: string,
   message?: string,
-  value?: TileValue,
+  metrics?: TileMetrics,
   build?: TileBuild
 }
+
+export default TileState
